fix: add `hashtag` parameter for Facebook
Allow to specify `hashtag` to pick the one tag that should be passed on to the Facebook share dialog. Sometimes, the developer might want to have the Twitter hashtags in a certain order and the one hashtag that should be passed on to Facebook is not the first in this specific order. If `hashtag` is not specified it defaults to (the first hashtag in) `hashtags`